requirements.txt
asgiref==3.8.1
attrs==23.2.0
cfgv==3.4.0
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
django-redis==5.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
factory-boy==3.3.0
Faker==26.0.0
filelock==3.15.4
hiredis==2.3.2
identify==2.6.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
nodeenv==1.9.1
platformdirs==4.2.2
pre-commit==3.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
redis==5.0.7
referencing==0.35.1
rpds-py==0.19.0
six==1.16.0
sqlparse==0.5.0
uritemplate==4.1.1
virtualenv==20.26.3
